legal-rag-scraper-deployment/requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
urllib3<2.0
tenacity>=8.2.3
websocket-client
zstandard
tiktoken>=0.5.0
azure-search-documents
azure-identity
azure-core
openai
python-dotenv
